Create Dockerfile
Dockerfile to build and extract nginx binaries from source. Uses Ubuntu 16.04 as the base image to build on as using older OS images is considered a healthier practice in managing libgc incompatibilities. This one is compiled with OpenSSL, a module that is also built from source in the same Dockerfile